Time: 37 ms (59.09%), Space: 96 MB (78.5%) - LeetHub

+++ 2316-count-unreachable-pairs-of-nodes-in-an-undirected-graph/2316-count-unreachable-pairs-of-nodes-in-an-undirected-graph.java
@@ -33,12 +33,12 @@ class Solution {
         }
         long res = 0L;
 
-        Map<Integer,Integer> map = new HashMap<>();
+        Map<Long,Long> map = new HashMap<>();
 
         for(int i = 0 ; i < n ; i++) 
         {
             int f = dsu.find(i);
-            map.put(f,map.getOrDefault(f,0)+1);
+            map.put((long) f,(long)map.getOrDefault((long)f,0L)+1);
         }
